+++ doc/changelog.rst
@@ -12,6 +12,7 @@ Unreleased
 Core
 ~~~~
 
+- Exclude images from product `get_public_media`
 - Add parameter to `PriceDisplayFilter` to specify tax display mode
 - Add soft deletion of categories
 - Add support to sell products after stock is zero